+++ lib/index.js
@@ -97,7 +97,7 @@ const initCore = (myID, params) => {
 };
 
 const reduceInit = (state, action) => {
-  const {
+  let {
     myID,
     myClass,
     doMe,
@@ -105,6 +105,7 @@ const reduceInit = (state, action) => {
     doParent,
     state: theState
   } = action;
+  theState = theState || {};
   let me = {
     id: myID,
     state: theState,
