+++ package.json
@@ -28,7 +28,8 @@
     "compile:lib": "babel src --out-dir lib --extensions '.tsx'",
     "build:lib": "npm run clean:lib && npm run compile:lib",
     "build": "npm run build:lib",
-    "test": "jest"
+    "test": "jest",
+    "coverage": "npx jest --coverage --silent --colors"
   },
   "license": "MIT",
   "peerDependencies": {
@@ -49,6 +50,8 @@
     "@babel/preset-modules": "^0.1.5",
     "@babel/preset-react": "^7.16.7",
     "@babel/preset-typescript": "^7.16.7",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.1.1",
     "@types/jest": "^27.4.1",
     "@types/react": "^18.0.5",
     "@types/react-dom": "^18.0.1",
@@ -59,7 +62,8 @@
     "babel-preset-env-modules": "^1.0.1",
     "cross-env": "^7.0.3",
     "eslint": "^8.13.0",
-    "jest": "^26.6.3",
+    "jest": "^27.5.1",
+    "jest-cli": "^27.5.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "ts-jest": "^27.1.4",
